Prepare deploy to Fly.io

+++ priv/repo/migrations/20191205022423_add_trades_table.exs
@@ -5,8 +5,8 @@ defmodule Oliya.Repo.Migrations.AddTradesTable do
     create table(:trades) do
       add(:venue, :string, null: false)
       add(:symbol, :string, null: false)
-      add(:price, :float, null: false)
-      add(:volume, :float, null: false)
+      add(:price, :decimal, null: false)
+      add(:volume, :decimal, null: false)
       add(:timestamp, :utc_datetime_usec, null: false)
       add(:side, :boolean, null: false)
       add(:venue_trade_id, :string)
